+++ src/commands/CurrentMonthPublications.tsx
@@ -35,9 +35,9 @@ export default function CurrentMonthPublications() {
     const searchDate = selectedDate.toLowerCase();
     return Object.entries(mangaList).reduce((filteredMangasByDate, [publicationDate, mangas]) => {
       const filteredMangas = mangas.filter(
-        ({ name, publicationDate }) =>
-          (!searchText || name.toLowerCase().includes(searchTitle)) &&
-          (!selectedDate || publicationDate.toLowerCase().includes(searchDate))
+        ({ name, publicationDate, editorial }) =>
+        (!searchText || (name + editorial).toLowerCase().includes(searchTitle)) &&
+        (!selectedDate || publicationDate.toLowerCase().includes(searchDate))
       );
       if (filteredMangas.length > 0) {
         filteredMangasByDate[publicationDate] = filteredMangas;